integrate into execution witness
