+++ src/ethereum_spec_tools/evm_tools/loaders/fork_loader.py
@@ -318,10 +318,6 @@ class ForkLoad:
         """track_ancestor_access function of the fork."""
         return self._module("state_tracker").track_ancestor_access
 
-    @property
-    def get_witness_ancestors(self) -> Any:
-        """get_witness_ancestors function of the fork."""
-        return self._module("state_tracker").get_witness_ancestors
 
     @property
     def State(self) -> Any:
